LPS-201959 Modify test to reveal issue
To properly test the import for FragmentEntryLink, the original
FragmentEntryLink and it's associated FragmentEntry should first
be removed, since the logic for importing a FragmentEntryLink
relies on the associated FragmentEntry to be available at the
time of import to set the proper FragmentEntryId value.

This would avoid the import accidentally working, due to the
original FragmentEntryLink or FragmentEntry being made available to
be referenced during the import.